refactor: move and rename collection related files into one collections/ folder
