+++ lib/app/layouts/conversation_view/widgets/message/attachment/collections/collection_attachment_card.dart
